+++ src/window/settingsdialog.cpp
@@ -7,7 +7,11 @@
 QString SettingsDialog::num_sections = QStringLiteral("maestro/num_sections");
 QString SettingsDialog::pixel_padding = QStringLiteral("interface/padding");
 QString SettingsDialog::pixel_shape = QStringLiteral("interface/shape");
+QString SettingsDialog::output_devices = QStringLiteral("serial/outputs");
+QString SettingsDialog::output_enabled = QStringLiteral("enabled");
+QString SettingsDialog::output_name = QStringLiteral("port");
 QString SettingsDialog::refresh_rate = QStringLiteral("maestro/refresh");
+QString SettingsDialog::screen_option = QStringLiteral("screen");
 QString SettingsDialog::serial_enabled = QStringLiteral("serial/enabled");
 QString SettingsDialog::serial_port = QStringLiteral("serial/port");
 QString SettingsDialog::virtual_device_option = QStringLiteral("Simulated Device");
@@ -18,35 +22,39 @@ SettingsDialog::SettingsDialog(QWidget *parent) : QDialog(parent), ui(new Ui::Se
 	ui->setupUi(this);
 
 	// Interface settings
-	ui->paddingComboBox->setCurrentIndex(settings_.value(pixel_padding).toInt());
-	ui->pixelShapeComboBox->setCurrentIndex(settings_.value(pixel_shape).toInt());
+	ui->paddingComboBox->setCurrentIndex(settings_.value(pixel_padding, 0).toInt());	// Default to no padding
+	ui->pixelShapeComboBox->setCurrentIndex(settings_.value(pixel_shape, 1).toInt());	// Default to square pixels
 
 	// Maestro settings
-	ui->numSectionsSpinBox->setValue(settings_.value(num_sections, 1).toInt());
-	ui->refreshSpinBox->setValue(settings_.value(refresh_rate, 50).toInt());
+	ui->numSectionsSpinBox->setValue(settings_.value(num_sections, 1).toInt());			// Default to 1 Section
+	ui->refreshSpinBox->setValue(settings_.value(refresh_rate, 50).toInt());			// Default to 50 ms
 
-	// Serial settings
-	ui->serialCheckBox->setChecked(settings_.value(serial_enabled).toBool());
-	on_serialCheckBox_toggled(ui->serialCheckBox->isChecked());
-
-	// Port settings
-	// Add option to select virtual serial device
-	ui->serialPortComboBox->addItem(SettingsDialog::virtual_device_option);
-
-	// Add actual serial devices
+	// Add serial devices to output selection box
 	QList<QSerialPortInfo> ports = QSerialPortInfo::availablePorts();
 	for (QSerialPortInfo port : ports) {
-		ui->serialPortComboBox->addItem(port.systemLocation());
+		QListWidgetItem* item = new QListWidgetItem(port.systemLocation());
+		item->setFlags(item->flags() | Qt::ItemIsUserCheckable);
+		item->setCheckState(Qt::Unchecked);
+		ui->outputListWidget->addItem(item);
 	}
 
-	// Only enable port dropdown if checked
-	ui->serialPortComboBox->setEnabled(ui->serialCheckBox->isChecked());
+	// Check to see if each entry in the output selection box has already been configured.
+	int num_devices = settings_.beginReadArray(output_devices);
+	for (int list_item = 0; list_item < ui->outputListWidget->count(); list_item++) {
+		for (int device = 0; device < num_devices; device++) {
+			settings_.setArrayIndex(device);
+			QListWidgetItem* item = ui->outputListWidget->item(list_item);
+			if (settings_.contains(output_name) && settings_.value(output_name).toString().compare(item->text(), Qt::CaseInsensitive) == 0) {
+				item->setCheckState((Qt::CheckState)settings_.value(output_enabled).toInt());
+			}
 
-	// Pre-select saved port
-	QString port_name = settings_.value(serial_port).toString();
-	ui->serialPortComboBox->setCurrentText(port_name);
-
-	set_simulated_device_options_enabled(ui->serialCheckBox->isChecked() && port_name.contains(virtual_device_option));
+			// If this is the Simulated Device, toggle the simulated device controls
+			if (item->text().compare(virtual_device_option, Qt::CaseInsensitive) == 0) {
+				set_simulated_device_options_enabled(item->checkState() == Qt::Checked);
+			}
+		}
+	}
+	settings_.endArray();
 }
 
 void SettingsDialog::on_buttonBox_accepted() {
@@ -54,9 +62,15 @@ void SettingsDialog::on_buttonBox_accepted() {
 	settings_.setValue(refresh_rate, ui->refreshSpinBox->value());
 	settings_.setValue(num_sections, ui->numSectionsSpinBox->value());
 
-	// Save serial settings
-	settings_.setValue(serial_enabled, ui->serialCheckBox->isChecked());
-	settings_.setValue(serial_port, ui->serialPortComboBox->currentText());
+	// Save selected output devices
+	settings_.beginWriteArray(output_devices);
+	for (int i = 0; i < ui->outputListWidget->count(); i++) {
+		settings_.setArrayIndex(i);
+		QListWidgetItem* item = ui->outputListWidget->item(i);
+		settings_.setValue(output_name, item->text());
+		settings_.setValue(output_enabled, item->checkState());
+	}
+	settings_.endArray();
 
 	// Save interface settings
 	settings_.setValue(pixel_padding, ui->paddingComboBox->currentIndex());
@@ -67,23 +81,20 @@ void SettingsDialog::on_buttonBox_accepted() {
 	settings_.setValue(virtual_device_height, ui->simulatedHeightSpinBox->value());
 }
 
-void SettingsDialog::on_serialCheckBox_toggled(bool checked) {
-	ui->serialPortComboBox->setEnabled(checked);
-
-	set_simulated_device_options_enabled(checked && ui->serialPortComboBox->currentText().contains(virtual_device_option));
+void SettingsDialog::on_outputListWidget_itemChanged(QListWidgetItem *item) {
+	// For Simulated Devices, set the serial device size options enabled/disabled.
+	if (item->text().compare(virtual_device_option, Qt::CaseInsensitive) == 0) {
+		set_simulated_device_options_enabled(item->checkState() == Qt::Checked);
+	}
 }
 
 void SettingsDialog::set_simulated_device_options_enabled(bool enabled) {
-	ui->simulatedDeviceHeightLabel->setEnabled(enabled);
 	ui->simulatedDeviceWidthLabel->setEnabled(enabled);
-	ui->simulatedHeightSpinBox->setEnabled(enabled);
+	ui->simulatedDeviceHeightLabel->setEnabled(enabled);
 	ui->simulatedWidthSpinBox->setEnabled(enabled);
+	ui->simulatedHeightSpinBox->setEnabled(enabled);
 }
 
 SettingsDialog::~SettingsDialog() {
 	delete ui;
-}
-
-void SettingsDialog::on_serialPortComboBox_currentTextChanged(const QString &arg1) {
-	set_simulated_device_options_enabled(arg1.contains(virtual_device_option));
 }